Renamed test_deps.ts to deps.ts
