+++ tests/cli/create_app_test.ts
@@ -5,7 +5,7 @@
  * This is only for some tests
  */
 
-import { Rhum } from "../test_deps.ts";
+import { Rhum } from "../deps.ts";
 import { red, green } from "../../deps.ts";
 import members from "../members.ts";
 const tmpDirName = "tmp-dir-for-testing-create-app";
