+++ tests/deps.ts
@@ -1,0 +1,6 @@
+export { Rhum } from "https://deno.land/x/rhum@v1.0.0/mod.ts";
+export * as path from "https://deno.land/std@v0.60.0/path/mod.ts";
+export {
+  MultipartReader,
+  isFormFile,
+} from "https://deno.land/std@v0.60.0/mime/multipart.ts";
