+++ tests/integration/app_3000_resources/coffee_resource_test.ts
@@ -1,5 +1,5 @@
 import members from "../../members.ts";
-import { Rhum } from "../../test_deps.ts";
+import { Rhum } from "../../deps.ts";
 
 Rhum.testPlan("coffee_resource_test.ts", () => {
   Rhum.testSuite("/coffee", () => {
