+++ tests/integration/app_3000_resources/request_accepts_resource_test.ts
@@ -1,5 +1,5 @@
 import members from "../../members.ts";
-import { Rhum } from "../../test_deps.ts";
+import { Rhum } from "../../deps.js";
 
 Rhum.testPlan("request_accepts_resource_test.ts", () => {
   Rhum.testSuite("/request-accepts", () => {
