+++ tests/integration/app_3001_views/template_engine_resource_test.ts
@@ -1,5 +1,5 @@
 import members from "../../members.ts";
-import { Rhum } from "../../test_deps.ts";
+import { Rhum } from "../../deps.js";
 
 function getExpected() {
   if (Deno.build.os == "windows") {
