--- tests/test_deps.ts
@@ -1,6 +1,0 @@
-export { Rhum } from "https://deno.land/x/rhum@v1.0.0/mod.ts";
-export * as path from "https://deno.land/std@v0.59.0/path/mod.ts";
-export {
-  MultipartReader,
-  isFormFile,
-} from "https://deno.land/std@v0.59.0/mime/multipart.ts";
